Move fieldAlign commands to a new class called FieldAlignment
